Initial Linux testing work
git-svn-id: https://mumble.svn.sourceforge.net/svnroot/mumble/trunk@335 05730e5d-ab1b-0410-a4ac-84af385074fa

+++ AudioInput.cpp
@@ -108,7 +108,7 @@ AudioInput::AudioInput()
 	psMic = new short[iFrameSize];
 	psSpeaker = new short[iFrameSize];
 	psClean = new short[iFrameSize];
-	pfY = new float[iFrameSize+1];
+	pfY = new int[iFrameSize+1];
 
 	bHasSpeaker = false;
